+++ pymc3/step_methods/hmc/base_hmc.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 
 import numpy as np
-
+import logging
 from pymc3.model import modelcontext, Point
 from pymc3.step_methods import arraystep
 from pymc3.step_methods.hmc import integration
@@ -10,16 +10,14 @@ from pymc3.tuning import guess_scaling
 from .quadpotential import quad_potential, QuadPotentialDiagAdapt
 from pymc3.step_methods import step_sizes
 from pymc3.backends.report import SamplerWarning, WarningType
+from pymc3.exceptions import SamplingError
+
+logger = logging.getLogger("pymc3")
+
+HMCStepData = namedtuple("HMCStepData", "end, accept_stat, divergence_info, stats")
 
 
-HMCStepData = namedtuple(
-    "HMCStepData",
-    "end, accept_stat, divergence_info, stats")
-
-
-DivergenceInfo = namedtuple(
-    'DivergenceInfo',
-    'message, exec_info, state')
+DivergenceInfo = namedtuple("DivergenceInfo", "message, exec_info, state")
 
 
 class BaseHMC(arraystep.GradientSharedStep):
@@ -27,12 +25,26 @@ class BaseHMC(arraystep.GradientSharedStep):
 
     default_blocked = True
 
-    def __init__(self, vars=None, scaling=None, step_scale=0.25, is_cov=False,
-                 model=None, blocked=True, potential=None,
-                 integrator="leapfrog", dtype=None, Emax=1000,
-                 target_accept=0.8, gamma=0.05, k=0.75, t0=10,
-                 adapt_step_size=True, step_rand=None,
-                 **theano_kwargs):
+    def __init__(
+        self,
+        vars=None,
+        scaling=None,
+        step_scale=0.25,
+        is_cov=False,
+        model=None,
+        blocked=True,
+        potential=None,
+        integrator="leapfrog",
+        dtype=None,
+        Emax=1000,
+        target_accept=0.8,
+        gamma=0.05,
+        k=0.75,
+        t0=10,
+        adapt_step_size=True,
+        step_rand=None,
+        **theano_kwargs
+    ):
         """Set up Hamiltonian samplers with common structures.
 
         Parameters
@@ -53,14 +65,15 @@ class BaseHMC(arraystep.GradientSharedStep):
             `energy`, and `random` methods.
         **theano_kwargs: passed to theano functions
         """
-        model = modelcontext(model)
+        self._model = modelcontext(model)
 
         if vars is None:
-            vars = model.cont_vars
+            vars = self._model.cont_vars
         vars = inputvars(vars)
 
-        super(BaseHMC, self).__init__(vars, blocked=blocked, model=model,
-                                      dtype=dtype, **theano_kwargs)
+        super(BaseHMC, self).__init__(
+            vars, blocked=blocked, model=model, dtype=dtype, **theano_kwargs
+        )
 
         self.adapt_step_size = adapt_step_size
         self.Emax = Emax
@@ -70,7 +83,8 @@ class BaseHMC(arraystep.GradientSharedStep):
         self.step_size = step_scale / (size ** 0.25)
         self.target_accept = target_accept
         self.step_adapt = step_sizes.DualAverageAdaptation(
-            self.step_size, target_accept, gamma, k, t0)
+            self.step_size, target_accept, gamma, k, t0
+        )
 
         self.tune = True
 
@@ -92,7 +106,8 @@ class BaseHMC(arraystep.GradientSharedStep):
             self.potential = quad_potential(scaling, is_cov)
 
         self.integrator = integration.CpuLeapfrogIntegrator(
-            self.potential, self._logp_dlogp_func)
+            self.potential, self._logp_dlogp_func
+        )
 
         self._step_rand = step_rand
         self._warnings = []
@@ -112,9 +127,26 @@ class BaseHMC(arraystep.GradientSharedStep):
         start = self.integrator.compute_state(q0, p0)
 
         if not np.isfinite(start.energy):
+            model = self._model
+            check_test_point = model.check_test_point()
+            error_logp = check_test_point.loc[
+                (np.abs(check_test_point) >= 1e20) | np.isnan(check_test_point)
+            ]
             self.potential.raise_ok(self._logp_dlogp_func._ordering.vmap)
-            raise ValueError('Bad initial energy: %s. The model '
-                             'might be misspecified.' % start.energy)
+            message_energy = (
+                "Bad initial energy, check any log probabilities that "
+                "are inf or -inf, nan or very small:\n{}".format(error_logp.to_string())
+            )
+            warning = SamplerWarning(
+                WarningType.BAD_ENERGY,
+                message_energy,
+                "critical",
+                self.iter_count,
+                None,
+                None,
+            )
+            self._warnings.append(warning)
+            raise SamplingError("Bad initial energy")
 
         adapt_step = self.tune and self.adapt_step_size
         step_size = self.step_adapt.current(adapt_step)
@@ -141,8 +173,8 @@ class BaseHMC(arraystep.GradientSharedStep):
                 else:
                     point = None
             warning = SamplerWarning(
-                kind, info.message, 'debug', self.iter_count,
-                info.exec_info, point)
+                kind, info.message, "debug", self.iter_count, info.exec_info, point
+            )
 
             self._warnings.append(warning)
 
@@ -150,10 +182,7 @@ class BaseHMC(arraystep.GradientSharedStep):
         if not self.tune:
             self._samples_after_tune += 1
 
-        stats = {
-            'tune': self.tune,
-            'diverging': bool(hmc_step.divergence_info),
-        }
+        stats = {"tune": self.tune, "diverging": bool(hmc_step.divergence_info)}
 
         stats.update(hmc_step.stats)
         stats.update(self.step_adapt.stats())
@@ -169,21 +198,28 @@ class BaseHMC(arraystep.GradientSharedStep):
         warnings = self._warnings[:]
 
         # Generate a global warning for divergences
-        message = ''
+        message = ""
         n_divs = self._num_divs_sample
         if n_divs and self._samples_after_tune == n_divs:
-            message = ('The chain contains only diverging samples. The model '
-                       'is probably misspecified.')
+            message = (
+                "The chain contains only diverging samples. The model "
+                "is probably misspecified."
+            )
         elif n_divs == 1:
-            message = ('There was 1 divergence after tuning. Increase '
-                       '`target_accept` or reparameterize.')
+            message = (
+                "There was 1 divergence after tuning. Increase "
+                "`target_accept` or reparameterize."
+            )
         elif n_divs > 1:
-            message = ('There were %s divergences after tuning. Increase '
-                       '`target_accept` or reparameterize.' % n_divs)
+            message = (
+                "There were %s divergences after tuning. Increase "
+                "`target_accept` or reparameterize." % n_divs
+            )
 
         if message:
             warning = SamplerWarning(
-                WarningType.DIVERGENCES, message, 'error', None, None, None)
+                WarningType.DIVERGENCES, message, "error", None, None, None
+            )
             warnings.append(warning)
 
         warnings.extend(self.step_adapt.warnings())